+++ R/spec-connection-disconnect.R
@@ -15,20 +15,6 @@ spec_connection_disconnect <- list(
     expect_invisible_true(dbDisconnect(con))
   },
 
-  #' An error is raised when attempting to disconnect from an already closed
-  disconnect_closed_connection = function(ctx) {
-    with_closed_connection({
-      expect_error(dbDisconnect(con))
-    })
-  },
-
-  #' or invalid connection.
-  disconnect_invalid_connection = function(ctx) {
-    with_invalid_connection({
-      expect_error(dbDisconnect(con))
-    })
-  },
-
   #' @section Specification:
   cannot_forget_disconnect = function(ctx) {
     expect_warning(gc(), NA)
@@ -38,12 +24,19 @@ spec_connection_disconnect <- list(
     expect_warning(gc())
   },
 
-  cannot_disconnect_twice = function(ctx) {
-    con <- connect(ctx)
-    dbDisconnect(con)
-    #' A warning is issued immediately when calling `dbDisconnect()` on an
-    #' already disconnected connection.
-    expect_warning(dbDisconnect(con))
+  #' A warning is issued immediately when calling `dbDisconnect()` on an
+  #' already disconnected
+  disconnect_closed_connection = function(ctx) {
+    with_closed_connection({
+      expect_warning(dbDisconnect(con))
+    })
+  },
+
+  #' or invalid connection.
+  disconnect_invalid_connection = function(ctx) {
+    with_invalid_connection({
+      expect_warning(dbDisconnect(con))
+    })
   },
 
   NULL